# api.yml
openapi: '3.0.4'
info:
  title: Sports Facilities Management System
  description: |-
    Esta es la definición de la API que se va a usar dentro del sistema de gestión de centros deportivos para poder hacer uso de todas las funciones que ofrece. Estas pueden ser separadas en función del tipo de usuario final que va a poder acceder a ellas:
    - Administradores: gestión de instalaciones, dispositivos y usuarios.
    - Clientes: gestión del perfil y reserva de instalaciones.
  contact:
    email: pabloquesadarojo@gmail.com
  version: '1.3.0'
servers:
  - url: https://api.server.test/
security:
  - BearerAuth: []
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Obtiene una lista de usuarios del sistema.
      description: Obtiene una lista de usuarios del sistema que puede estar filtrada en función de una serie de parámetros.
      operationId: getUsers
      x-security-roles:
        - SUPERADMIN
        - ADMIN
      parameters:
        - name: role
          in: query
          required: false
          description: Rol del usuario a filtrar.
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: id
          in: query
          required: false
          description: Identificador del usuario a filtrar.
          schema:
            type: number
        - name: name
          in: query
          required: false
          description: Nombre del usuario a filtrar.
          schema:
            type: string
        - name: surname
          in: query
          required: false
          description: Apellido del usuario a filtrar.
          schema:
            type: string
        - name: mail
          in: query
          required: false
          description: Correo electrónico del usuario a filtrar.
          schema:
            type: string
            format: email
        - name: phonePrefix
          in: query
          required: false
          description: Prefijo del número de teléfono del usuario a filtrar.
          schema:
            type: number
        - name: phoneNumber
          in: query
          required: false
          description: Número de teléfono del usuario a filtrar.
          schema:
            type: number
      responses:
        '200':
          description: Se devuelve correctamente la lista de usuarios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWithoutPassword'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - users
      summary: Crea un nuevo usuario en el sistema.
      description: Crea un nuevo usuario en el sistema, asignándole un identificador único y los datos necesarios.
      operationId: createUser
      x-security-roles:
        - SUPERADMIN
        - ANONYMOUS
      requestBody:
        description: Objeto que contiene la información del usuario a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserWithRole'
        required: true
      responses:
        '201':
          description: El usuario se ha creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithoutPassword'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      tags:
        - users
      summary: Obtiene la información de un usuario.
      description: |-
        Obtiene la información de un usuario concreto.
        - Los superadmisnistradores/administradores pueden obtener la información de cualquier usuario.
        - Los usuarios autenticados pueden obtener sólo su propia información.
      operationId: getUserInfo
      x-security-roles:
        - SUPERADMIN
        - ADMIN
        - USER
      responses:
        '200':
          description: Se devuelve correctamente la información del usuario pedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el usuario.
          content:
            application/json: {}
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - users
      summary: Actualiza la información de un usuario.
      description: |-
        Actualiza la información de un usuario concreto.
        - Los superadministradores/administradores pueden actualizar la información de cualquier usuario.
        - Los usuarios autenticados pueden actualizar sólo su propia información.
      operationId: updateUserInfo
      x-security-roles:
        - SUPERADMIN
        - ADMIN
        - USER
      requestBody:
        description: Objeto que contiene la información actualizada del usuario.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: La información del usuario se ha actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el usuario.
          content:
            application/json: {}
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - users
      summary: Elimina un usuario del sistema.
      description: |-
        Elimina un usuario del sistema, borrando toda su información asociada.
        - Los superadministradores/administradores pueden eliminar cualquier usuario.
        - Los usuarios autenticados pueden eliminar sólo su propia cuenta.
      operationId: deleteUser
      x-security-roles:
        - SUPERADMIN
        - ADMIN
        - USER
      responses:
        '204':
          description: El usuario se ha eliminado correctamente.
          content:
            application/json: {}
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el usuario.
          content:
            application/json: {}
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complexes:
    get:
      tags:
        - complexes
      summary: Obtiene una lista de complejos deportivos del sistema.
      description: Obtiene una lista de complejos deportivos del sistema que puede estar filtrada en función de una serie de parámetros.
      operationId: getComplexes
      x-security-roles:
        - ADMIN
        - USER
      responses:
        '200':
          description: Se devuelve correctamente la lista de complejos deportivos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplexInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - complexes
      summary: Crea un nuevo complejo deportivo en el sistema.
      description: Crea un nuevo complejo deportivo en el sistema, asignándole un identificador único y completando la información asociada a este.
      operationId: createComplex
      x-security-roles:
        - SUPERADMIN
      requestBody:
        description: Objeto que contiene la información del complejo deportivo a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplexInfo'
        required: true
      responses:
        '200':
          description: El complejo deportivo se ha creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplexInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complexes/{complexId}/reservations:
    parameters:
      - $ref: '#/components/parameters/complexId'
    get:
      tags:
        - reservations
      summary: Obtiene una lista de reservas de pista.
      description: |-
        Obtiene una lista de reservas de pista que puede estar filtrada en función de una serie de parámetros.
        - Los administradores pueden obtener todas las reservas.
        - Los usuarios autenticados pueden obtener sólo sus propias reservas.
      operationId: getReservations
      x-security-roles:
        - ADMIN
        - USER
      parameters:
        - name: userId
          in: query
          required: false
          description: Identificador del usuario para filtrar las reservas.
          schema:
            type: integer
        - name: courtId
          in: query
          required: false
          description: Identificador de la pista para filtrar las reservas.
          schema:
            type: integer
      responses:
        '200':
          description: Se devuelve correctamente la lista de reservas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se han encontrado reservas.
          content:
            application/json: {}
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - reservations
      summary: Crea una nueva reserva de pista.
      description: Crea una nueva reserva de pista, asignando un usuario y una pista específica para un horario determinado.
      operationId: createReservation
      x-security-roles:
        - ADMIN
        - USER
      requestBody:
        description: Objeto que contiene la información de la reserva a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
        required: true
      responses:
        '201':
          description: La reserva se ha creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: Código de estado HTTP.
        message:
          type: string
          description: Mensaje descriptivo del error.
        error:
          type: string
          description: Tipo de error (ej. Bad Request, Unauthorized).
...
    WeatherData:
      type: object
      properties:
        temperature:
          type: number
          description: Temperatura actual.
        precip_intensity_prev:
          type: number
          description: Precipitación registrada en la hora previa.
        precip_intensity_curr:
          type: number
          description: Precipitación actual.
        precip_probability_curr:
          type: integer
          description: Probabilidad de precipitación actual.
        precip_probability_next:
          type: integer
          description: Probabilidad de precipitación para la próxima hora.
      description: Datos meteorológicos obtenidos de la API de OpenMeteo.
...
    UserRole:
      type: string
      enum:
        - UNKNOWN
        - ADMIN
        - USER
        - DEVICE
    UserCredentials:
      type: object
      properties:
        mail:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 64
      required:
        - mail
        - password
    User:
      type: object
      properties:
        credentials:
          $ref: '#/components/schemas/UserCredentials'
        name:
          type: string
          minLength: 2
          maxLength: 50
        surname:
          type: string
          minLength: 2
          maxLength: 50
        phonePrefix:
          type: integer
          minimum: 1
          maximum: 999
        phoneNumber:
          type: integer
          minimum: 10000
      required:
        - credentials
        - name
        - phonePrefix
        - phoneNumber
    UserWithRole:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
        user:
          $ref: '#/components/schemas/User'
      required:
        - role
    UserWithoutPassword:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
        mail:
          type: string
          format: email
        name:
          type: string
        surname:
          type: string
        phonePrefix:
          type: integer
          minimum: 1
          maximum: 999
        phoneNumber:
          type: integer
          minimum: 10000
...
    TimeInfo:
      type: object
      properties:
        timeIni:
          type: string
          format: time
        timeEnd:
          type: string
          format: time
    ComplexInfo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        time:
          $ref: '#/components/schemas/TimeInfo'
        locLongitude:
          type: number
          format: double
        locLatitude:
          type: number
          format: double
        weather:
          $ref: '#/components/schemas/WeatherData'
...
    Reservation:
      type: object
      properties:
        courtId:
          type: integer
        time:
          $ref: '#/components/schemas/TimeInfo'
...
  parameters:
    complexId:
      name: complexId
      in: path
      required: true
      description: Identificador del sistema.
      schema:
        type: integer
...
    userId:
      name: userId
      in: path
      required: true
      description: Identificador del usuario.
      schema:
        type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceAuth:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Token de autenticación del dispositivo, utilizado para autenticar las peticiones de los dispositivos IoT.