# src/reservations/complex-reservations.controller.ts
import { Body, Controller, Get, Param, ParseIntPipe, Post, Query } from '@nestjs/common';
import { Roles } from 'src/auth/decorator';
import { Role } from 'src/auth/enums';
import { CreateReservationDto, GetReservationsDto } from './dto';
import { ReservationsService } from './reservations.service';

@Controller('complexes')
export class ComplexReservationsController {
  constructor(private reservationsService: ReservationsService) {}

  @Roles(Role.CLIENT, Role.ADMIN)
  @Get(':complexId/reservations')
  async getComplexReservations(
    @Param('complexId', ParseIntPipe) complexId: number,
    @Query() query: GetReservationsDto,
  ) {
    return this.reservationsService.getComplexReservations(complexId, query);
  }

  @Roles(Role.CLIENT, Role.ADMIN)
  @Post(':complexId/reservations')
  async createComplexReservation(
    @Param('complexId', ParseIntPipe) complexId: number,
    @Body() dto: CreateReservationDto,
  ) {
    return this.reservationsService.createReservation(complexId, dto);
  }
}